// frontend/portfolioERP-web/src/app/features/customers/models/create-customer-request.ts
export interface CreateCustomerRequest {
  customerCode: string;
  companyName: string;
  firstName: string | null;
  lastName: string | null;
  taxCode: string | null;
  vatNumber: string | null;
  email: string;
  phone: string | null;
  address: string | null;
  city: string | null;
  province: string | null;
  postalCode: string | null;
  country: string;
}